% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_admetlab.R
\name{admetlab_tab}
\alias{admetlab_tab}
\title{ADMETlab 3.0 Manager tab UI}
\usage{
admetlab_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Builds the ADMETlab 3.0 Manager tab of the ADMETShiny application.
}
\keyword{internal}
