% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{ADMETShiny application server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Shiny input, output and session objects.}
}
\description{
The server function backing the ADMETShiny Shiny application. It is not
meant to be called directly by end users; use \code{\link{run_app}} to
launch the app.
}
\keyword{internal}
