% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_customization.R
\name{apply_palette}
\alias{apply_palette}
\title{Apply a colour palette to a ggplot object}
\usage{
apply_palette(p, palette_name, data, color_col)
}
\arguments{
\item{p}{A ggplot object.}

\item{palette_name}{Character. Name of the palette (as returned by
\code{\link{palette_selector_ui}}).}

\item{data}{The data.frame used to build \code{p}.}

\item{color_col}{Character. Name of the column used for colouring, or
\code{"None"}.}
}
\value{
A ggplot object (possibly modified).
}
\description{
Modifies an existing ggplot object by adding a colour/fill scale matching
the selected palette, intelligently detecting whether the colour variable
is continuous or discrete.
}
\examples{
\dontrun{
library(ggplot2)
d <- data.frame(MW = c(300, 400, 500), group = c("A", "A", "B"))
p <- ggplot(d, aes(x = group, y = MW, color = group)) + geom_point()
p <- apply_palette(p, "Set1", d, "group")
}
}
