% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_deeppk.R
\name{deeppk_tab}
\alias{deeppk_tab}
\title{Deep-PK Manager tab UI}
\usage{
deeppk_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Builds the Deep-PK Manager tab of the ADMETShiny application.
}
\keyword{internal}
