% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_home.R
\name{home_tab}
\alias{home_tab}
\title{Home tab UI
Morrigan, i love you! Thank you my friend.
Builds the Home (landing) tab of the ADMETShiny application.}
\usage{
home_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Home tab UI
Morrigan, i love you! Thank you my friend.
Builds the Home (landing) tab of the ADMETShiny application.
}
\keyword{internal}
