% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_customization.R
\name{palette_selector_ui}
\alias{palette_selector_ui}
\title{Colour palette selector UI element}
\usage{
palette_selector_ui(id)
}
\arguments{
\item{id}{Character. Input id to use for the select input.}
}
\value{
A \code{shiny.tag} (select input).
}
\description{
A Shiny \code{selectInput} that lets the user choose a colour palette for
the plots produced by the application.
}
