% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling \code{rhs(lhs, ...)}.
}
\description{
Re-exports \code{\link[magrittr:pipe]{\%>\%}} from \pkg{magrittr} so that
the pipe can be used with \code{admetshiny::\%>\%} or after attaching the
package.
}
\keyword{internal}
