% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLogP}
\alias{plotLogP}
\title{LogP distribution histogram}
\usage{
plotLogP(data)
}
\arguments{
\item{data}{A data.frame containing a numeric \code{LogP} column.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
LogP distribution histogram
}
\examples{
\dontrun{
d <- data.frame(LogP = c(1, 2, 3))
plotLogP(d)
}
}
