% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMW}
\alias{plotMW}
\title{Molecular weight distribution histogram}
\usage{
plotMW(data)
}
\arguments{
\item{data}{A data.frame containing a numeric \code{MW} column.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Molecular weight distribution histogram
}
\examples{
\dontrun{
d <- data.frame(MW = c(300, 400, 500))
plotMW(d)
}
}
