% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTPSA}
\alias{plotTPSA}
\title{TPSA distribution histogram}
\usage{
plotTPSA(data)
}
\arguments{
\item{data}{A data.frame containing a numeric \code{TPSA} column.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
TPSA distribution histogram
}
\examples{
\dontrun{
d <- data.frame(TPSA = c(40, 80, 120))
plotTPSA(d)
}
}
