% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_swissadme.R
\name{swissadme_tab}
\alias{swissadme_tab}
\title{SwissADME Manager tab UI}
\usage{
swissadme_tab()
}
\value{
A \code{shiny.tabPanel}.
}
\description{
Builds the SwissADME Manager tab of the ADMETShiny application.
}
\keyword{internal}
